Create advanced_padder.py (#35)
* Create jump_touch_reward.py

a ball touch reward that only triggers when the agent's wheels aren't in contact with the floor adjust minimum ball height required for reward with 'min_height' as well as reward scaling with 'exp'

* Create advanced_padder.py

AdvancedObsPadder is essentially the default advanced obs observer class with 0 padding added on. It also comes with the added benefit of being able to "expand" to work natively with rocket learn framework or to work with sb3 in it's default configuration.